requirements.txt
pandas
requests
django
djangorestframework
mysqlclient
django-cors-headers
python-decouple
djangorestframework-simplejwt
drf-yasg
django_redis
drf-nested-routers
psycopg2
gunicorn
whitenoise
firebase_admin
Pillow
django-cors-headers
celery